pymor.parameters.interface.ParameterFunctional -> pymor.parameters.functionals.ParameterFunctional
