+++ src/pymor/operators/basic.py
@@ -10,7 +10,7 @@ from pymor.algorithms import genericsolvers
 from pymor.core.base import abstractmethod
 from pymor.core.exceptions import InversionError, LinAlgError
 from pymor.operators.interface import Operator
-from pymor.parameters.interfaces import ParameterFunctional
+from pymor.parameters.functionals import ParameterFunctional
 from pymor.vectorarrays.interface import VectorArray
 from pymor.vectorarrays.numpy import NumpyVectorSpace
